// src/API.ts
export type CreateWebMentionEventInput = {
  targetId: string,
  targetSlug: string,
  targetName: string,
  targetType?: string | null,
  createdAt?: string | null,
  updatedAt?: string | null,
  type: WebMentionType,
  secret: string,
  source: string,
  target: string,
  postType: string,
  postUrl: string,
  authorName?: string | null,
  authorPhoto?: string | null,
  authorUrl?: string | null,
  published: string,
  name?: string | null,
};

export enum WebMentionType {
  InReplyTo = "InReplyTo",
  LikeOf = "LikeOf",
  RepostOf = "RepostOf",
  BookmarkOf = "BookmarkOf",
  MentionOf = "MentionOf",
  RSVP = "RSVP",
}